fix: handle `iterable` response in scheduled task controller
The method `AbstractMessageHandler::getHandledMessages()` has
`iterable` as its return type, but the code in
`ScheduledTaskController` assumed that its result can be passed as the
second argument to `in_array()`. This causes errors when using the
"start manually" button in the scheduled tasks manager UI.

Change this to convert the result into an array before passsing it into
`in_array()`.